Added Filter.js file
