+++ src/components/Filter.js
@@ -1,0 +1,17 @@
+import { useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { fetchCountries } from '../redux/country/countrySlice';
+
+const Filter = () => {
+  const dispatch = useDispatch();
+  const countries = useSelector((state) => state.countries.countries);
+
+  useEffect(() => {
+    dispatch(fetchCountries());
+  }, [dispatch]);
+
+  const filteredCountries = countries.filter((country) => country.region === 'Africa');
+  return filteredCountries;
+};
+
+export default Filter;
